backend/requirements.txt
bcrypt==4.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
distlib==0.3.4
filelock==3.6.0
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
netifaces==0.11.0
packaging==23.0
pbr==5.8.1
platformdirs==2.5.1
PyJWT==2.6.0
PyMySQL==1.0.2
requests==2.27.1
six==1.16.0
SQLAlchemy==2.0.4
stevedore==3.5.0
typing_extensions==4.5.0
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
Werkzeug==2.2.3
zipp==3.15.0
